Added float32

+++ stable-diffusion/custom/data/mnist_dataset.py
@@ -62,8 +62,9 @@ class MNISTDataset(Dataset):
         if self.transform:
             image = self.transform(image)
 
+#float 32?
         #shift image values from [0,1] into [-1,1]
-        image = (image * 2.0) - 1.0
+        image = ((image * 2.0) - 1.0).to(torch.float32)
 
         output["image"] = image
         output["caption"] = label
